Replacing normalizer/validator with interfaces.
